Splitting sync and task drivers into iree/hal/drivers/. Progress on #4298.
