+++ runtime/src/iree/hal/drivers/local_sync/CMakeLists.txt
@@ -1,6 +1,6 @@
 ################################################################################
 # Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
-# runtime/src/iree/hal/drivers/dylib_sync/registration/BUILD                   #
+# runtime/src/iree/hal/drivers/local_sync/BUILD                                #
 #                                                                              #
 # Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
 # CMake-only content.                                                          #
@@ -12,19 +12,28 @@ iree_add_all_subdirs()
 
 iree_cc_library(
   NAME
-    registration
+    sync_driver
   HDRS
-    "driver_module.h"
+    "sync_device.h"
+    "sync_driver.h"
+    "sync_event.h"
+    "sync_semaphore.h"
   SRCS
-    "driver_module.c"
+    "sync_device.c"
+    "sync_driver.c"
+    "sync_event.c"
+    "sync_semaphore.c"
   DEPS
     iree::base
+    iree::base::core_headers
+    iree::base::internal
+    iree::base::internal::arena
+    iree::base::internal::synchronization
+    iree::base::tracing
     iree::hal
-    iree::hal::drivers::local_sync::sync_driver
     iree::hal::local
-    iree::hal::local::loaders::embedded_library_loader
-  DEFINES
-    "IREE_HAL_HAVE_DYLIB_SYNC_DRIVER_MODULE=1"
+    iree::hal::utils::buffer_transfer
+    iree::hal::utils::semaphore_base
   PUBLIC
 )
