+++ runtime/src/iree/hal/drivers/local_sync/sync_driver.c
@@ -4,7 +4,7 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "iree/hal/local/sync_driver.h"
+#include "iree/hal/drivers/local_sync/sync_driver.h"
 
 #include <stddef.h>
 #include <string.h>
