+++ runtime/src/iree/hal/drivers/local_task/CMakeLists.txt
@@ -1,0 +1,52 @@
+################################################################################
+# Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
+# runtime/src/iree/hal/drivers/local_task/BUILD                                #
+#                                                                              #
+# Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
+# CMake-only content.                                                          #
+#                                                                              #
+# To disable autogeneration for this file entirely, delete this header.        #
+################################################################################
+
+iree_add_all_subdirs()
+
+iree_cc_library(
+  NAME
+    task_driver
+  HDRS
+    "task_command_buffer.h"
+    "task_device.h"
+    "task_driver.h"
+    "task_event.h"
+    "task_queue.h"
+    "task_queue_state.h"
+    "task_semaphore.h"
+  SRCS
+    "task_command_buffer.c"
+    "task_device.c"
+    "task_driver.c"
+    "task_event.c"
+    "task_queue.c"
+    "task_queue_state.c"
+    "task_semaphore.c"
+  DEPS
+    iree::base
+    iree::base::core_headers
+    iree::base::internal
+    iree::base::internal::arena
+    iree::base::internal::event_pool
+    iree::base::internal::synchronization
+    iree::base::internal::wait_handle
+    iree::base::tracing
+    iree::hal
+    iree::hal::local
+    iree::hal::local::executable_environment
+    iree::hal::local::executable_library
+    iree::hal::utils::buffer_transfer
+    iree::hal::utils::resource_set
+    iree::hal::utils::semaphore_base
+    iree::task
+  PUBLIC
+)
+
+### BAZEL_TO_CMAKE_PRESERVES_ALL_CONTENT_BELOW_THIS_LINE ###
