+++ runtime/src/iree/hal/drivers/local_task/task_command_buffer.c
@@ -4,7 +4,7 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "iree/hal/local/task_command_buffer.h"
+#include "iree/hal/drivers/local_task/task_command_buffer.h"
 
 #include <stdbool.h>
 #include <stddef.h>
