+++ runtime/src/iree/hal/drivers/local_task/task_command_buffer.h
@@ -4,13 +4,13 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#ifndef IREE_HAL_LOCAL_TASK_COMMAND_BUFFER_H_
-#define IREE_HAL_LOCAL_TASK_COMMAND_BUFFER_H_
+#ifndef IREE_HAL_DRIVERS_LOCAL_TASK_TASK_COMMAND_BUFFER_H_
+#define IREE_HAL_DRIVERS_LOCAL_TASK_TASK_COMMAND_BUFFER_H_
 
 #include "iree/base/api.h"
 #include "iree/base/internal/arena.h"
 #include "iree/hal/api.h"
-#include "iree/hal/local/task_queue_state.h"
+#include "iree/hal/drivers/local_task/task_queue_state.h"
 #include "iree/task/scope.h"
 #include "iree/task/task.h"
 
@@ -55,4 +55,4 @@ iree_status_t iree_hal_task_command_buffer_issue(
 }  // extern "C"
 #endif  // __cplusplus
 
-#endif  // IREE_HAL_LOCAL_TASK_COMMAND_BUFFER_H_
+#endif  // IREE_HAL_DRIVERS_LOCAL_TASK_TASK_COMMAND_BUFFER_H_
