+++ runtime/src/iree/hal/drivers/local_task/task_queue_state.h
@@ -4,8 +4,8 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#ifndef IREE_HAL_LOCAL_TASK_QUEUE_STATE_H_
-#define IREE_HAL_LOCAL_TASK_QUEUE_STATE_H_
+#ifndef IREE_HAL_DRIVERS_LOCAL_TASK_TASK_QUEUE_STATE_H_
+#define IREE_HAL_DRIVERS_LOCAL_TASK_TASK_QUEUE_STATE_H_
 
 #include "iree/base/api.h"
 #include "iree/base/internal/atomics.h"
@@ -38,4 +38,4 @@ void iree_hal_task_queue_state_deinitialize(
 }  // extern "C"
 #endif  // __cplusplus
 
-#endif  // IREE_HAL_LOCAL_TASK_QUEUE_STATE_H_
+#endif  // IREE_HAL_DRIVERS_LOCAL_TASK_TASK_QUEUE_STATE_H_
