+++ runtime/src/iree/hal/local/BUILD
@@ -111,87 +111,3 @@ iree_runtime_cc_library(
         "//runtime/src/iree/hal",
     ],
 )
-
-iree_runtime_cc_library(
-    name = "sync_driver",
-    srcs = [
-        "sync_device.c",
-        "sync_driver.c",
-        "sync_event.c",
-        "sync_semaphore.c",
-    ],
-    hdrs = [
-        "sync_device.h",
-        "sync_driver.h",
-        "sync_event.h",
-        "sync_semaphore.h",
-    ],
-    deps = [
-        ":local",
-        "//runtime/src/iree/base",
-        "//runtime/src/iree/base:core_headers",
-        "//runtime/src/iree/base:tracing",
-        "//runtime/src/iree/base/internal",
-        "//runtime/src/iree/base/internal:arena",
-        "//runtime/src/iree/base/internal:synchronization",
-        "//runtime/src/iree/hal",
-        "//runtime/src/iree/hal/utils:buffer_transfer",
-        "//runtime/src/iree/hal/utils:semaphore_base",
-    ],
-)
-
-#===------------------------------------------------------------------------===#
-# Thread dependent packages
-#===------------------------------------------------------------------------===#
-
-iree_cmake_extra_content(
-    content = """
-# task_driver is used by asynchronuous drivers.
-# TODO(scotttodd): refactor this - code depending on threading should be
-#   possible to declare in the build system but conditionally link in
-if(NOT EMSCRIPTEN AND NOT (IREE_HAL_DRIVER_DYLIB OR IREE_HAL_DRIVER_VMVX))
-  return()
-endif()
-""",
-    inline = True,
-)
-
-iree_runtime_cc_library(
-    name = "task_driver",
-    srcs = [
-        "task_command_buffer.c",
-        "task_device.c",
-        "task_driver.c",
-        "task_event.c",
-        "task_queue.c",
-        "task_queue_state.c",
-        "task_semaphore.c",
-    ],
-    hdrs = [
-        "task_command_buffer.h",
-        "task_device.h",
-        "task_driver.h",
-        "task_event.h",
-        "task_queue.h",
-        "task_queue_state.h",
-        "task_semaphore.h",
-    ],
-    deps = [
-        ":executable_environment",
-        ":executable_library",
-        ":local",
-        "//runtime/src/iree/base",
-        "//runtime/src/iree/base:core_headers",
-        "//runtime/src/iree/base:tracing",
-        "//runtime/src/iree/base/internal",
-        "//runtime/src/iree/base/internal:arena",
-        "//runtime/src/iree/base/internal:event_pool",
-        "//runtime/src/iree/base/internal:synchronization",
-        "//runtime/src/iree/base/internal:wait_handle",
-        "//runtime/src/iree/hal",
-        "//runtime/src/iree/hal/utils:buffer_transfer",
-        "//runtime/src/iree/hal/utils:resource_set",
-        "//runtime/src/iree/hal/utils:semaphore_base",
-        "//runtime/src/iree/task",
-    ],
-)
