+++ runtime/src/iree/hal/local/CMakeLists.txt
@@ -104,77 +104,4 @@ iree_cc_library(
   PUBLIC
 )
 
-iree_cc_library(
-  NAME
-    sync_driver
-  HDRS
-    "sync_device.h"
-    "sync_driver.h"
-    "sync_event.h"
-    "sync_semaphore.h"
-  SRCS
-    "sync_device.c"
-    "sync_driver.c"
-    "sync_event.c"
-    "sync_semaphore.c"
-  DEPS
-    ::local
-    iree::base
-    iree::base::core_headers
-    iree::base::internal
-    iree::base::internal::arena
-    iree::base::internal::synchronization
-    iree::base::tracing
-    iree::hal
-    iree::hal::utils::buffer_transfer
-    iree::hal::utils::semaphore_base
-  PUBLIC
-)
-
-# task_driver is used by asynchronuous drivers.
-# TODO(scotttodd): refactor this - code depending on threading should be
-#   possible to declare in the build system but conditionally link in
-if(NOT EMSCRIPTEN AND NOT (IREE_HAL_DRIVER_DYLIB OR IREE_HAL_DRIVER_VMVX))
-  return()
-endif()
-
-iree_cc_library(
-  NAME
-    task_driver
-  HDRS
-    "task_command_buffer.h"
-    "task_device.h"
-    "task_driver.h"
-    "task_event.h"
-    "task_queue.h"
-    "task_queue_state.h"
-    "task_semaphore.h"
-  SRCS
-    "task_command_buffer.c"
-    "task_device.c"
-    "task_driver.c"
-    "task_event.c"
-    "task_queue.c"
-    "task_queue_state.c"
-    "task_semaphore.c"
-  DEPS
-    ::executable_environment
-    ::executable_library
-    ::local
-    iree::base
-    iree::base::core_headers
-    iree::base::internal
-    iree::base::internal::arena
-    iree::base::internal::event_pool
-    iree::base::internal::synchronization
-    iree::base::internal::wait_handle
-    iree::base::tracing
-    iree::hal
-    iree::hal::utils::buffer_transfer
-    iree::hal::utils::resource_set
-    iree::hal::utils::semaphore_base
-    iree::task
-  PUBLIC
-)
-
 ### BAZEL_TO_CMAKE_PRESERVES_ALL_CONTENT_BELOW_THIS_LINE ###
