+++ samples/simple_embedding/BUILD
@@ -32,8 +32,8 @@ cc_binary(
         ":simple_embedding_test_bytecode_module_vmvx_c",
         "//runtime/src/iree/base",
         "//runtime/src/iree/hal",
+        "//runtime/src/iree/hal/drivers/local_sync:sync_driver",
         "//runtime/src/iree/hal/local",
-        "//runtime/src/iree/hal/local:sync_driver",
         "//runtime/src/iree/hal/local/loaders:vmvx_module_loader",
         "//runtime/src/iree/modules/hal",
         "//runtime/src/iree/vm",
@@ -86,8 +86,8 @@ cc_binary(
         ":simple_embedding_test_bytecode_module_dylib_x86_64_c",
         "//runtime/src/iree/base",
         "//runtime/src/iree/hal",
+        "//runtime/src/iree/hal/drivers/local_sync:sync_driver",
         "//runtime/src/iree/hal/local",
-        "//runtime/src/iree/hal/local:sync_driver",
         "//runtime/src/iree/hal/local/loaders:embedded_library_loader",
         "//runtime/src/iree/modules/hal",
         "//runtime/src/iree/vm",
@@ -201,8 +201,8 @@ cc_binary(
         ":simple_embedding_test_bytecode_module_dylib_x86_64_c",
         "//runtime/src/iree/base",
         "//runtime/src/iree/hal",
+        "//runtime/src/iree/hal/drivers/local_task:task_driver",
         "//runtime/src/iree/hal/local",
-        "//runtime/src/iree/hal/local:task_driver",
         "//runtime/src/iree/hal/local/loaders:embedded_library_loader",
         "//runtime/src/iree/modules/hal",
         "//runtime/src/iree/task:api",
