+++ samples/simple_embedding/device_dylib.c
@@ -10,9 +10,9 @@
 
 #include "iree/base/api.h"
 #include "iree/hal/api.h"
+#include "iree/hal/drivers/local_task/task_device.h"
 #include "iree/hal/local/executable_loader.h"
 #include "iree/hal/local/loaders/embedded_library_loader.h"
-#include "iree/hal/local/task_device.h"
 #include "iree/task/api.h"
 
 // Compiled module embedded here to avoid file IO:
