+++ samples/static_library/static_library_demo.c
@@ -7,8 +7,8 @@
 // A example of static library loading in IREE. See the README.md for more info.
 // Note: this demo requires artifacts from iree-compile before it will run.
 
+#include "iree/hal/drivers/local_sync/sync_device.h"
 #include "iree/hal/local/loaders/static_library_loader.h"
-#include "iree/hal/local/sync_device.h"
 #include "iree/modules/hal/module.h"
 #include "iree/runtime/api.h"
